store, core: Create `store-test` crate, refactor tests to use it
