+++ core/Cargo.toml
@@ -21,8 +21,6 @@ serde_yaml = "0.7"
 ipfs-api = "0.5.0-alpha2"
 graph-mock = { path = "../mock" }
 walkdir = "2.2.5"
-diesel = { version = "1.3.2", features = ["postgres", "serde_json", "numeric", "r2d2"] }
-graph-store-postgres = { path = "../store/postgres" }
-lazy_static = "1.2.0"
+test-store = { path = "../store/test-store" }
 hex = "0.3.2"
 graphql-parser = "0.2.1"
