+++ store/postgres/Cargo.toml
@@ -1,6 +1,7 @@
 [package]
 name = "graph-store-postgres"
 version = "0.5.0"
+edition = "2018"
 
 [dependencies]
 bigdecimal = "0.0.14"
@@ -17,5 +18,6 @@ serde = "1.0"
 uuid = { version = "0.7.2", features = ["v4"] }
 
 [dev-dependencies]
+test-store = { path = "../test-store" }
 lazy_static = "1.1"
 hex = "0.3.2"
