+++ store/test-store/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "test-store"
+version = "0.1.0"
+authors = ["Leonardo Yvens <leoyvens@gmail.com>"]
+edition = "2018"
+description = "Provides static store instance for tests."
+
+[dependencies]
+graph = { path = "../../graph" }
+graph-store-postgres = { path = "../postgres" }
+lazy_static = "1.1"
